requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.piptools
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
beautifulsoup4==4.12.3
    # via -r requirements.piptools
blinker==1.8.2
    # via flask
boto3==1.34.103
    # via -r requirements.piptools
botocore==1.34.103
    # via
    #   -r requirements.piptools
    #   boto3
    #   s3transfer
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
distro==1.9.0
    # via openai
flask==3.0.3
    # via
    #   -r requirements.piptools
    #   flask-cors
flask-cors==4.0.1
    # via -r requirements.piptools
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
openai==1.28.1
    # via -r requirements.piptools
packaging==24.0
    # via pytest
pluggy==1.5.0
    # via pytest
psycopg2==2.9.9
    # via -r requirements.piptools
pydantic==2.7.1
    # via openai
#pydantic-core==2.18.2
    # via pydantic
pytest==8.2.0
    # via -r requirements.piptools
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.0.1
    # via -r requirements.piptools
pytube==15.0.0
    # via -r requirements.piptools
requests==2.31.0
    # via -r requirements.piptools
s3transfer==0.10.1
    # via boto3
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
tqdm==4.66.4
    # via openai
#typing-extensions==4.11.0
    # via
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via
    #   -r requirements.piptools
    #   botocore
    #   requests
werkzeug==3.0.3
    # via flask
